Resolve compiler warnings in engine code
- Fix signedness comparison warnings (-Wsign-compare)
  - globals.cc: use size_t for loop iterating NB_POLICY_VALUES()
  - benchmark.cc: use size_t for loop iterating NB_INPUT_VALUES()
  - training_data_writer.cc: use size_t for loop and channel variable
- Fix deprecated enum-float arithmetic warnings
  - evaluate.cpp: cast PawnValueEg to int before division
  - uci.cpp: cast PawnValueEg to int before division
- Fix unused variable warning in test_move_gen.cc
  - Remove unused moves_a_initial variable

+++ engine/src/benchmark.cc
@@ -17,7 +17,7 @@ void benchmark_inference(Engine& engine, int iterations) {
     float* piB = new float[SearchParams::BATCH_SIZE * NB_POLICY_VALUES()];
     
     // Initialize with random data
-    for (int i = 0; i < SearchParams::BATCH_SIZE * NB_INPUT_VALUES(); i++) {
+    for (size_t i = 0; i < SearchParams::BATCH_SIZE * NB_INPUT_VALUES(); i++) {
         obs[i] = static_cast<float>(rand()) / RAND_MAX;
     }
